% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{calc_manual_countries}
\alias{calc_manual_countries}
\title{Find and write countries which need manual processing}
\usage{
calc_manual_countries()
}
\description{
This functions reads the file with errors for the current date in the folder \href{https://github.com/finddx/FINDCov19TrackerData/tree/master/issues}{issues},
and generates a template for the current day in \href{https://github.com/finddx/FINDCov19TrackerData/tree/master/manual/need-processing}{manual/need-processing/}.

This process updates the file \code{-need-manual-processing.csv} \emph{only} for the current day.
To see if there are errors for past dates check the folder \href{https://github.com/finddx/FINDCov19TrackerData/tree/master/issues}{issues}

For the automated countries with error \code{test_cumulative} is set to 0 in the templates \href{https://github.com/finddx/FINDCov19TrackerData/tree/master/manual/need-processing}{manual/need-processing/}.
}
