% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{check_country}
\alias{check_country}
\title{Check and process country specific tasks}
\usage{
check_country(dots, ...)
}
\description{
This function ensures that valid values are returned for countries.
}
