% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-data.R
\name{check_jhu_data}
\alias{check_jhu_data}
\title{Perform Consistency Checks}
\usage{
check_jhu_data(original, modified)
}
\description{
Compares original and preprocessed JHU datasets to ensure
consistency.
}
