% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-data.R
\name{create_shiny_data}
\alias{create_shiny_data}
\title{Create Input Data for Shiny Application}
\usage{
create_shiny_data()
}
\description{
Reads coronavirus_cases.csv (cases) and coronavirus_tests.csv (tests)
files from \verb{processed/} directory.
Writes \code{processed/data_shiny.csv}.
}
