% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{FINDCov19Tracker-package}
\alias{FINDCov19Tracker}
\alias{FINDCov19Tracker-package}
\title{FINDCov19Tracker: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dsbbfinddx/FINDCov19Tracker}
}

}
\author{
\strong{Maintainer}: Patrick Schratz \email{patrick.schratz@gmail.com} (\href{https://orcid.org/0000-0003-0748-6624}{ORCID})

}
