% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-data.R
\name{get_test_data}
\alias{get_test_data}
\title{Get and combine test data from different sources}
\usage{
get_test_data(days = 1, write = TRUE)
}
\arguments{
\item{days}{days to combine. It should always be bigger than 0.}

\item{write}{if csv files should be written. Default TRUE.}
}
\description{
\strong{Input:} Daily test data scraped via Selenium and "R fetch functions" from
\href{https://github.com/dsbbfinddx/FINDCov19TrackerData/tree/master/automated/fetch}{\code{automated/fetch}} and \href{https://github.com/dsbbfinddx/FINDCov19TrackerData/tree/master/automated/selenium}{\code{automated/selenium}} directories.

\strong{Output:}
\itemize{
\item \code{automated-tests.json}
\item \code{countries-error.csv}
}

which are then deployed by CI to the \href{https://github.com/dsbbfinddx/FINDCov19TrackerData/tree/master/automated/merged}{\verb{automated/merged/}} directory in the \code{dsbbfinddx/FINDCov19TrackerData} repo.
}
