% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-data.R
\name{preprocess_jhu_data}
\alias{preprocess_jhu_data}
\title{Preprocess JHU data
Processes JHU input data}
\usage{
preprocess_jhu_data(input_df, tag)
}
\description{
Preprocess JHU data
Processes JHU input data
}
