% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_test_data}
\alias{preprocess_test_data}
\title{Scrape test data from all countries}
\usage{
preprocess_test_data()
}
\description{
Rowwise processing of countries test data reports with support for any file
type (csv, pdf, xlsx, etc.)
}
