% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-data.R
\name{process_jhu_data}
\alias{process_jhu_data}
\title{Update John Hopkins University data}
\usage{
process_jhu_data()
}
\value{
Writes \code{processed/coronavirus_cases.csv} and \code{processed/jhu_data.csv}.
}
\description{
Updates the JHU cases data and saves them in coronavirus_cases.csv
file in the processed folder

Edward Parker, London School of Hygiene & Tropical Medicine, March 2019.
Data extracted from Johns Hopkins data obtained from following Github
repository https://github.com/CSSEGISandData/COVID-19
}
