% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-data.R
\name{process_test_data}
\alias{process_test_data}
\title{Postprocess tests data for input into Shiny App}
\usage{
process_test_data()
}
\value{
Writes \code{coronavirus_tests.csv}
}
\description{
During the process, \code{coronavirus_cases.csv} (case data) is read and used.
This file is created/updated by \code{process_jhu_data()} which needs to be run
before.
}
