% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{calc_manual_countries}
\alias{calc_manual_countries}
\title{Find and write countries which need manual processing}
\usage{
calc_manual_countries()
}
\description{
This functions reads the list of \href{https://github.com/dsbbfinddx/FINDCov19TrackerData/blob/master/manual/countries-urls.csv}{automated countries} via
Selenium and R fetch functions and calculates which countries still need
manual processing.
}
