% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{fetch_test_data}
\alias{fetch_test_data}
\title{Scrape test data from all countries via R}
\usage{
fetch_test_data()
}
\description{
Rowwise processing of countries test data reports with support for various
file type (csv, pdf, xlsx, etc.).
Used for all countries which cannot be scraped via Selenium.
}
