% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-data.R
\name{get_daily_test_data}
\alias{get_daily_test_data}
\title{Get Test Data}
\usage{
get_daily_test_data()
}
\description{
Gets daily test data scraped via Selenium and R fetch functions from
FINDCov19TrackerData data branch.
The data is cleaned and combined and written to \code{automated-tests.json}
which is then deployed by CI to \verb{automated/merged/} directory within the
FINDCov19TrackerData \code{data} branch.
}
