% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-data.R
\name{process_test_data}
\alias{process_test_data}
\title{Postprocess tests data for input into Shiny App}
\usage{
process_test_data()
}
\description{
Creates \code{coronavirus_cases.csv}, based on \code{coronavirus_cases_new.csv}.
\code{coronavirus_cases_new.csv} combines different sources (Selenium, fetch, and manual),
and might have NA or negative values for new_test_corrected.
}
\details{
This file is created/updated by \code{process_jhu_data()} which needs to be run
before.
\strong{Input:} test data combined from different sources (Selenium, fetch, and manual)\href{https://github.com/dsbbfinddx/FINDCov19TrackerData/blob/master/automated/coronavirus_tests_new.csv}{\code{automated/coronavirus_tests_new.csv}}.

#'   \strong{Output:}
\itemize{
\item \code{coronavirus_cases.csv}
}

which is then deployed by CI to the \href{https://github.com/dsbbfinddx/FINDCov19TrackerData/tree/master/processed}{\verb{processed/}} directory in the \code{dsbbfinddx/FINDCov19TrackerData} repo.
}
